Add handling of CLICOLOR, MINETEST_COLOR environment variables (#7553)
For CLICOLOR, see https://bixense.com/clicolors/ (CLICOLOR_FORCE not
implemented yet, not sure if we should use this one at all, it sounds
like a source of confusion.

For MINETEST_COLOR, the same options as `--color` are used.

They are evaluated in this order (overriding each other):

1. CLICOLOR
2. MINETEST_COLOR
3. commandline arguments

so in order of increasing specificity.