Reassure previous nil behaviour for tiles
The mod `dumpnodes` which is included with minetestmapper has broken
with 5.6.0 when run with `bike` mod, due to changes in 18fbc0394 which
defines a tiles table for all nodes by default, where previously a nil
was returned if a nil was given. Restore previous behaviour so mods do
not break.