refacto(Server): more intelligent packet handler
Server packet handler now have the RemoteClient as a parameter and
handlers have a minimum session state prerequisite.

Why this change ? Nearly all packets are re-lookuping the RemoteClient
(based on session state, we keep that), then just send the parameter to all handlers and use it.
It also prevent a lock on the session to perform the lookup, which is now unnecessary as we have the session to handle.

Also we don't check the session state in majority of cases, or it's
implicit because of the getClient call, but getClient returns nullptr is
session is not in right state whereas an existing peer_id may be valid.
It's a pretty illogical way to handle the getter as it can lead to
coding errors (rare, we do in majority things in CS_Active state) and
it hides wrong things sent by clients (maybe rogues), or catch them up
very late in the code path.

Drop ClientInterface::getClientState, it's not needed anymore, it was a very low used helper, now we have the client we can directly check it

For the performance optimization it should be a bit low, except when you
have hundreds of client sending thousands of commands (that reduce the
locking to zero), then a very rare case. For the code usage security
it's clearly higher than before as now you need to set the required
state for packets.