Explain what a float range is
This was not documented anywhere in this documentation so I asked in the Discord. Thanks to @LoneWolfHT this was able to be figured out. But I shouldn't have even had to ask that question because it should have just been right there in that portion of the documentation, as it's not referred to anywhere else in the entire api docs.