formspec model camera position control
Fixes a bug with the formspec model element where the camera's orbit position
would change when the same form was opened multiple times.

Does so by adding API to control where the camera's orbit is, as offset from the model's origin,
and clamping controls for the distance the camera is from the orbit position.