Move behavior out of CAO mgr into other classes
This adds a Shootline class that does the selectionbox calculations
instead of the client ActiveObjectMgr doing them directly. It also
moves part of the logic to ClientActiveObject. This has greatly reduced
the size of the client ActiveObjectMgr class, and most of what's left is
object management behavior. In computer science jargon, we could say
the codebase is more cohesive.