Changed all core::unicode::ustring to std::wstring
I changed all the instances of core::unicode::ustring to std::wstring.
This was not too hard as this data structure is only used in two files.
Since IrrlichtMT does not use core::unicode::ustring anywhere except the
header in which it is defined (or anything else defined in that header),
we will be able to remove the include/irrUString.h file from the
IrrlichtMT repository, removing the last non-STL container from
IrrlichtMT.