Make safeWriteToFile more likely to succeed on Windows.
Use win32 APIs to write the temporary file before copying to the final
destination. Because we've observed the final file being empty, we
suspect that std::ostream::flush is not flushing.

Also add a test for it. It seemed some things in the `fs` namespace
were already tested in test_filepath, so I renamed that to test_filesys
and added the test there.

Change-Id: Idb817c3b41c7f4c81805f937e6bf467d85a4f124