Return to the main menu if a shader compilation fails
Before this change, if the shaders are broken, only an error message is shown and the player enters the world nonetheless, where he/she sees broken graphics.
For a shader developer, being thrown back to the main menu can be more convenient.

To throw the player back to the main menu, we add an exception.
Since BaseException is only catched if NDEBUG is set, we add a new exception dedicated to shaders.