Fix disappearing and reappearing node changes of dig and placement predictions
If a player has a slow connection to the server and they successively places nodes in the same mapblock,
the predictions can be temporarily undone if an outdated mapblock is received from the server.
This is especially problematic on servers with a Skyblock game.
A similar problem happens when digging nodes;
in this case nodes temporarily reappear.

Example; client side:
1. The player places a node at pos1. The prediction makes it appear immediately.
2. The player places a node at pos2, which is in the same mapblock. The prediction makes it appear immediately.
3. block1 arrives. It is integrated into the client-side map and undoes the prediction at pos2.
4. block2 arrives. It is integrated into the client-side map and redoes the predicted change at pos2.

Example; server side:
1. An interaction at pos1 arrives. The server modifies the map and sends the modified mapblock block1 to the client.
2. An interaction at pos2 arrives. The server modifies the map and sends the modified mapblock block2 to the client.

To fix (or noticeably mitigate) the problem,
we change the client so that it reapplies placement and dig predictions when a mapblock arrives while the sent interact packet is not yet acknowledged.

If, for some reason, an updated mapblock arrives even if the sent packet with interact information has not yet been acknowledged, ghost nodes can appear now.
This can happen if the server sends an updated mapblock before sending the acknowledgement for the arrival of the packet with interact information.
TODO: make it robust against packet reordering

TODO: I don't know if my understanding of the problem is correct