Translation updater: Add comment propagation
The comment propagation (--propagate-comment or -c) propagates comments from the template file into the translation files.
The comments from the template file are written above the comments from the translation file itself.
There is a simple check to make sure that the comments from the template file are not propagated twice.