use SDL2 for headless rendering
Xvfb is not needed if we use OpenGL + SDL2.
This makes the code much simpler (no need to start a separate Xvfb
process), it works without any additonial setup on our dev machine
(large-bass), and it runs at nearly 60FPS.

Also some clean up of the examples to let it work with locally built
minetest binary.

Fixes: https://github.com/Astera-org/minetest/issues/32