Fix snow man test texture coordinates
The current texture coordinates specified in the test did not use sufficient precision.
It is unfortunately common for software to round away the last bit when displaying floats.
This issue was previously masked by Irrlicht using a float comparison with tolerance for vectors,
which was changed in 3452857 to an exact float comparison for unrelated reasons.
Manual inspection of the buffer reveals that the values produced by the reader are correct.