add_area_node_boxes: Cap the maximum invalid nodes to avoid memory exhaustion
When the client trying to calculate a very high velocity (like the current speed
limit 5000), the client stuck in `add_area_node_boxes` trying to add all nodes
along and causing a memory exhaustion. The commit caps the function to abort
when invalid nodes exceeded 16 MiB to prevent such occasions.