Remove a workaround
At some point I tested with SDL_HINT_MOUSE_TOUCH_EVENTS=1 to be able to test
the touch controls with a mouse on desktop.
That resulted in the touch controls working, but mouse events arriving twice in GUIs:
- once directly via mouse
- once via mouse -> touch (converted by SDL) -> mouse (converted by Minetest,
  regular conversion that's necessary for menus to work on touchscreens at all)
This resulted in a crash here because the code had more side effects iirc. Now
it seems fine, and having a workaround for this specific case is questionable anyway.