Delete default `aabbox3d` default constructor
The default box is a 2x2x2 box centered around the origin, but the
default constructor conventionally creates an "empty" object. This
unexpected default behavior has caused bugs and confusion. We delete
the constructor so that any usage of the default constructor will cause
a compile-time error, ensuring that we have updated all uses of the API.
For now, we preserve exactly the old behavior, initializing all boxes as
the default constructor would have done it.