Silence failing raycast unit test
The cause for the test failure is an edge case bug
in the raycast implementation (perfectly diagonal raycasts).

This is done by switching to a continuous random distribution
which makes it extremely unlikely that the buggy edge case occurs.

Additionally, devtest unit test failures now print their random seed
to be easier to reproduce in the future.